refactor db; use EFCore.NamingConventions nuget, remove redundant code, allow cors *
