--- BookApi/BookApi.http
@@ -1,6 +1,0 @@
-@BookApi_HostAddress = http://localhost:5145
-
-GET {{BookApi_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
